VAVFA-8986: move covid status to separate component
