+++ src/applications/facility-locator/components/search-results-items/VaFacilityResult.jsx
@@ -8,6 +8,7 @@ import { recordResultClickEvents } from '../../utils/analytics';
 import { LocationType, OperatingStatus, Covid19Vaccine } from '../../constants';
 import LocationAddress from './common/LocationAddress';
 import LocationOperationStatus from './common/LocationOperationStatus';
+import LocationCovidStatus from './common/LocationCovidStatus';
 import LocationDistance from './common/LocationDistance';
 import Covid19Alert from './common/Covid19Alert';
 
@@ -50,6 +51,11 @@ const VaFacilityResult = ({
           operatingStatus.code !== OperatingStatus.NORMAL && (
             <LocationOperationStatus operatingStatus={operatingStatus} />
           )}
+        {operatingStatus && (
+          <LocationCovidStatus
+            supplementalStatus={operatingStatus.supplementalStatus}
+          />
+        )}
         <LocationAddress location={location} />
         <LocationDirectionsLink location={location} from="SearchResult" />
         {isCovid19Search && <Covid19Alert />}
