+++ src/applications/facility-locator/components/search-results-items/common/LocationOperationStatus.jsx
@@ -1,32 +1,9 @@
 /* eslint-disable jsx-a11y/no-noninteractive-tabindex */
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
-import createCommonStore from 'platform/startup/store';
-import { connectFeatureToggle } from 'platform/utilities/feature-toggles';
-import { FeatureToggleReducer } from 'platform/site-wide/feature-toggles/reducers';
-import { toggleValues } from 'platform/site-wide/feature-toggles/selectors';
 import { OperatingStatus } from '../../../constants';
 
-const store = createCommonStore({
-  FeatureToggleReducer,
-});
-
 const LocationOperationStatus = ({ operatingStatus }) => {
-  const [showVamcAlert, setShowVamcAlert] = useState(true);
-
-  useEffect(() => {
-    connectFeatureToggle(store.dispatch);
-    store.subscribe(() => {
-      const flags = toggleValues(store.getState());
-      if (
-        flags?.showExpandableVamcAlert === 'undefined' ||
-        flags?.showExpandableVamcAlert === false
-      ) {
-        setShowVamcAlert(false);
-      }
-    });
-  }, []);
-
   if (!operatingStatus || operatingStatus.code === 'NORMAL') {
     return <></>;
   }
@@ -53,40 +30,19 @@ const LocationOperationStatus = ({ operatingStatus }) => {
   const { operationStatusTitle, alertClass } = display[operatingStatus.code];
 
   return (
-    <div>
-      <va-alert
-        background-only
-        show-icon
-        status={alertClass}
-        visible
-        data-testid={`${operatingStatus.code.toLowerCase()}-message`}
-        class="vads-u-margin-y--2"
-      >
-        <div tabIndex={0}>
-          <span className="sr-only">Alert: </span>
-          {operationStatusTitle}
-        </div>
-      </va-alert>
-
-      {operatingStatus.supplementalStatus.map(status => {
-        if (status.id.includes('COVID') && showVamcAlert) {
-          return (
-            <va-alert
-              background-only
-              show-icon
-              status={alertClass}
-              visible
-              data-testid={`${operatingStatus.code.toLowerCase()}-message`}
-              class="vads-u-margin-y--2"
-            >
-              <div tabIndex={0}>{status.label}</div>
-            </va-alert>
-          );
-        }
-
-        return null;
-      })}
-    </div>
+    <va-alert
+      background-only
+      show-icon
+      status={alertClass}
+      visible
+      data-testid={`${operatingStatus.code.toLowerCase()}-message`}
+      class="vads-u-margin-y--2"
+    >
+      <div tabIndex={0}>
+        <span className="sr-only">Alert: </span>
+        {operationStatusTitle}
+      </div>
+    </va-alert>
   );
 };
